+++ CHANGELOG.md
@@ -1,12 +1,24 @@
+<<<<<<< HEAD
 ## [0.0.3] - May 22nd, 2018.
 
 * Change ``vibrate({Duration duration : _DEFAULT_VIBRATION_DURATION})`` to accept optional parameter ``duration`` for Android devices
 * Dart 2.0 compatibility
 
 ## [0.0.2] - January 26th, 2018.
+=======
+## [0.0.4] - August 22nd, 2018
+
+* add compatibility for Dart 2 sdk in pubspec.yaml
+
+## [0.0.3] - July 10th, 2018
+
+* haptic Feedback for iOS and Android
+
+## [0.0.2] - January 26th, 2018
+>>>>>>> 60e02647f66a1e73ecbd1fe294bf4d8992ed64be
 
 * added ``vibrateWithPauses(Iterable<Duration> pauses)``
 
-## [0.0.1] - January 24th, 2018.
+## [0.0.1] - January 24th, 2018
 
 * vibrate() and canVibrate() implemented on iOS and Android